Charts: rename the context directory to chart-instance-context
With the aliases gone, every import read `ChartInstanceContext` from a
path spelled `single-chart-context`. That mismatch is the thing a
reader actually trips on, and it is the last trace of the old name.

The directory and both files now match what they hold.
