+++ projects/js-packages/charts/src/charts/private/chart-instance-context/chart-instance-context.tsx
