+++ projects/js-packages/charts/src/charts/private/chart-instance-context/index.ts
@@ -1,0 +1,3 @@
+export { ChartInstanceContext } from './chart-instance-context';
+export { useChartInstanceContext } from './use-chart-instance-context';
+export type { ChartInstanceContextValue, ChartInstanceRef } from './chart-instance-context';
