+++ projects/js-packages/charts/src/charts/private/chart-instance-context/use-chart-instance-context.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ChartInstanceContext, type ChartInstanceContextValue } from './single-chart-context';
+import { ChartInstanceContext, type ChartInstanceContextValue } from './chart-instance-context';
 
 export const useChartInstanceContext = (): ChartInstanceContextValue => {
 	const context = useContext( ChartInstanceContext );
