--- projects/js-packages/charts/src/charts/private/single-chart-context/index.ts
@@ -1,3 +1,0 @@
-export { ChartInstanceContext } from './single-chart-context';
-export { useChartInstanceContext } from './use-single-chart-context';
-export type { ChartInstanceContextValue, ChartInstanceRef } from './single-chart-context';
