+++ projects/js-packages/charts/src/charts/private/x-zoom/test/x-zoom.test.tsx
@@ -2,7 +2,7 @@ import { act, render, renderHook, screen, waitFor } from '@testing-library/react
 import userEvent from '@testing-library/user-event';
 import { useCallback, useRef, useState } from 'react';
 import { useXZoom, ZoomResetButton } from '../index';
-import type { ChartInstanceRef } from '../../single-chart-context';
+import type { ChartInstanceRef } from '../../chart-instance-context';
 import type { EventHandlerParams } from '@visx/xychart';
 import type { KeyboardEvent as ReactKeyboardEvent } from 'react';
